[internal] - remove redundant cast byte to byte

+++ libse/ManagedBitmap.cs
@@ -146,10 +146,10 @@ namespace Nikse.SubtitleEdit.Core
         private static void WriteColor(Stream stream, Color c)
         {
             byte[] buffer = new byte[4];
-            buffer[0] = (byte)c.A;
-            buffer[1] = (byte)c.R;
-            buffer[2] = (byte)c.G;
-            buffer[3] = (byte)c.B;
+            buffer[0] = c.A;
+            buffer[1] = c.R;
+            buffer[2] = c.G;
+            buffer[3] = c.B;
             stream.Write(buffer, 0, buffer.Length);
         }
